Adds verbose option to PyBullet connection
Adds a `verbose` parameter to the `PyBulletClient.connect()` and `PyBulletClient.disconnect()` methods, allowing users to control the verbosity of the PyBullet connection.
This helps in debugging and avoids potential errors related to stdout redirection.

Changes the `PyBulletClient._handle_concavity()` to avoid the `redirect_stdout` context manager to avoid the mysterious `WinError 6: The handle is invalid.` error.